Third revision - Playable
-In this version-
The player can choose width, height and cellsize of grid.
The player can place a start, a goal and multiple platforms on the grid.
The text platforms placed converts into normal platforms that the player character can walk on.
Player movement system is now fully functional.
The player now starts on the spawn-platform.