The New Fourth rendition - Color and Location Fix
-In this version-
The player can choose width, height and cellsize of grid.
The player can place a start, a goal and multiple platforms on the grid.
The text platforms placed converts into normal platforms that the player character can walk on.
Player movement system is now fully functional.
The player now starts on the spawn-platform. (Fixed 4th rendition)
The platforms are color coded as to show the logic behind it.
The player is constrained to a location before getting placed on the startpoint.