Sixteenth rendition - Changes Travelscript
-In this version-
The travelscript is now toggle:able through the bool variable "sexyteleport".

This will be used to transfer the player from the hub to the dungeon, when the dungeon is complete the gate in the boss-room will open and the player can go to the boss-level. When the final boss is defeated, the player goes back to the hub.